Excluded Quant op from constant folding in cleanup and added transformation to fold transposes into Quant op to cleanup.
